task-runner/requirements.txt
openai-agents>=0.0.7
mcp>=1.0.0
pydantic>=2.0.0
